backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
psycopg[binary]==3.17.2
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
requests==2.31.0
pyotp==2.9.0
python-multipart==0.0.6
cors==1.0.1
promethus-client==0.19.0
pycryptodome==3.19.0
structlog==24.1.0
argon2-cffi==23.1.0
